+++ changelog.md
@@ -26,6 +26,8 @@ the feature `pallet-session-compat`.
 
 ## Added
 
+
+* `sign-tx` command to `smart-contracts` commands for signing transactions
 * `sign-block-producer-metadata` command to `cli-commands` for signing block producer metadata upsert message
 * `db-sync-sqlx` crate containing Rust types representing Cardano primitives present in postgress tables populated by Db-Sync